mediunp

+++ src/webl/shaders/shaderFirePlane.ts
@@ -426,7 +426,7 @@ export function GetShaderSourceFireVisualizerPS() {
         /* glsl */ `#version 300 es
 	
 	precision highp float;
-	precision highp sampler2D;
+	precision mediump sampler2D;
 
 	layout(location = 0) out vec4 OutFirePlane;
